moved most of client to imports/; changed a bit of the states
