--- client/admin/views/admin.ng.jade
@@ -1,6 +1,0 @@
-.admin
-  ul
-    li
-      a(ui-sref="admin.dhm-parse") dondehaymisa.com parsed output
-    li
-      quieroiramisa.com parsed output
