+++ client/index.html
@@ -1,0 +1,6 @@
+<head>
+  <base href="/">
+</head>
+<body>
+  <ui-view></ui-view>
+</body>
