+++ client/main.js
@@ -1,5 +1,7 @@
-
 import { Meteor } from 'meteor/meteor';
+
+import '../imports/ui/components/parroquias/parroquias';
+import '../imports/ui/routes';
 
 function onReady() {
   angular.bootstrap(document, [
