--- client/parroquias/views/parroquias.ng.jade
@@ -1,5 +1,0 @@
-div 
-  h1 Parroquias:
-  ul
-    li(ng-repeat="parroquia in psc.parroquias") 
-      a(ui-sref='parroquia({id: parroquia._id})') {{parroquia.name}}
