+++ imports/ui/components/admin/admin.ng.jade
@@ -1,0 +1,6 @@
+.admin
+  ul
+    li
+      a(ui-sref="misas.admin.dhm-parse") dondehaymisa.com parsed output
+    li
+      quieroiramisa.com parsed output
