+++ imports/ui/components/parroquias/parroquias.ng.jade
@@ -1,0 +1,10 @@
+div 
+  h1 Parroquias:
+  ul
+    li
+      a(ui-sref="misas") Home
+    li
+      a(ui-sref="misas.parroquias.search") Search Parroquias
+    li
+      a(ui-sref="misas.admin") Admin
+  ui-view
