+++ imports/ui/routes.js
@@ -1,35 +1,48 @@
+import './components/parroquias/parroquias';
+import './components/parroquia/parroquia';
+import './components/parroquias/parroquias.search';
+import './components/admin/sources/DHM/DHM-parse';
+
 angular.module('parroquias').config(function($urlRouterProvider, $stateProvider, $locationProvider) {
   $locationProvider.html5Mode(true);
-  $stateProvider.state('parroquias', {
-    url: '/parroquias/',
-    template: '<parroquias></parroquias>'
-  }).state('parroquias.search', {
-    url: 'search',
+  $stateProvider.state('misas', {
+    url: '/',
     views: {
       '@': {
+        template: '<parroquias></parroquias>'
+      }
+    }
+  }).state('misas.parroquias', {
+    abstract: true,
+    url: 'parroquias/',
+    template: '<ui-view/>'
+  }).state('misas.parroquias.search', {
+    url: 'search',
+    views: {
+      '@misas': {
         template: '<parroquias-search></parroquias-search>'
       }
     }
-  }).state('parroquia', {
-    url: '/parroquia/{id}',
+  }).state('misas.parroquia', {
+    url: 'parroquia/{id}',
     template: '<parroquia id="id"></parroquia>',
     controller: function($scope, $stateParams) {
       $scope.id = $stateParams.id;
       return console.log("parroquia state loaded");
     }
-  }).state('admin', {
-    url: '/admin/',
-    templateUrl: 'client/admin/views/admin.html',
+  }).state('misas.admin', {
+    url: 'admin/',
+    templateUrl: 'imports/ui/components/admin/admin.html',
     controller: function($scope) {
       return console.log("admin");
     }
-  }).state('admin.dhm-parse', {
+  }).state('misas.admin.dhm-parse', {
     url: 'dhm-parse/',
     controller: function($scope) {
       return console.log("dhm parsing");
     },
     views: {
-      '@': {
+      '@misas': {
         template: '<dhm-parse></dhm-parse>'
       }
     }
