--- parroquias.html
@@ -1,19 +1,0 @@
-<head>
-  <base href="/">
-</head>
-<body>
-  <div ng-app="parroquias" ng-strict-di="">
-    <ul>
-      <li>
-        <a ui-sref="parroquias">Home</a>
-      </li>
-      <li>
-        <a ui-sref="parroquias.search">Search</a>
-      </li>
-      <li>
-        <a ui-sref="admin">Admin</a>
-      </li>
-    </ul>
-    <div ui-view=""></div>
-  </div>
-</body>
