--- parroquias.js
@@ -1,8 +1,0 @@
-if (Meteor.isClient) {
-}
-
-if (Meteor.isServer) {
-  Meteor.startup(function () {
-    // code to run on server at startup
-  });
-}
